Move StatementSource to separate file
